chore: clean up implicit conversions

+++ src/config.cpp
@@ -775,7 +775,7 @@ namespace config {
     if (val.size() >= 2 && val.substr(0, 2) == "0x"sv) {
       input = util::from_hex<int>(val.substr(2));
     } else {
-      input = util::from_view(val);
+      input = (int) util::from_view(val);
     }
 
     vars.erase(it);
@@ -979,7 +979,7 @@ namespace config {
       if (val.size() >= 2 && val.substr(0, 2) == "0x"sv) {
         tmp = util::from_hex<int>(val.substr(2));
       } else {
-        tmp = util::from_view(val);
+        tmp = (int) util::from_view(val);
       }
       input.emplace_back(tmp);
     }
